+++ feincms3/templatetags/feincms3_pages.py
@@ -83,7 +83,8 @@ def is_descendant_of(node1, node2, include_self=False):
 
 
 @register.simple_tag
-def menu(menu, *, level=0, depth=1, **kwargs):
+# def menu(menu, *, level=0, depth=1, **kwargs):
+def menu(menu, level=0, depth=1, **kwargs):
     """menu(menu, level=0, depth=1, **kwargs)
     This tag expects the ``page`` variable to contain the page we're on
     currently. The active pages are fetched using ``.objects.active()`` and